+++ Adapter/ShopwareAdapter/ServiceBus/CommandHandler/Product/HandleProductCommandHandler.php
@@ -190,6 +190,7 @@ class HandleProductCommandHandler implements CommandHandlerInterface
         }
 
         $this->attributeHelper->addFieldAsAttribute($product, 'technicalDescription');
+        $this->attributeHelper->addFieldAsAttribute($product, 'ageRestriction');
 
         $this->attributeDataPersister->saveProductDetailAttributes(
             $productModel->getMainDetail(),